backend/requirements.txt
django>=5.2,<5.3
djangorestframework>=3.17
django-cors-headers>=4.9
Pillow>=12.0
python-dotenv>=1.0
anthropic>=0.28
celery>=5.4
redis>=5.0
